Clear cached report when child pipeline is updated
When a pipeline has child pipelines
and one of the child pipelines have been
updated, we need to also invalidate the
coverage report on the root pipeline.

This commit changes the key used to determine
the freshness of the data by using the
latest `updated_at` of the pipelines
in the hierarchy of pipelines,
isntead of only using the root pipeline's
`updated_at`.