Use dangers to remind user to check migrations
Because of limited capacity, the migration testing pipeline is still
manual only. As a result, I'm adding a reminder to the danger message
so that folks kick off the testing pipeline.
Also deduplicates database files and fixes docs review for database docs

Relates to https://gitlab.com/gitlab-org/gitlab/issues/328367