Fix issue list for closed issues without closed_at
Prevents a 500 in the dashboard issue list and fixes wrong timestamp
being used in vue issue list when a closed issue does not have a
closed_at value.

This can happen when an issue is imported from GitHub or for old issues
that were closed before we added the closed_at column.