Add new custom headers when a new destination is created
It will first create the destination and then if there are no errors it
will loop through all the headers and create them for the new
destination. If the headers save correctly, then the success is shown.
If the headers error, then the created destination is deleted and the
error shown.

Updated the docs and specs to match the new feature.