Reduce performance bar expiry to 5 minutes
Otherwise we'd use the default of 30 minutes. We generally don't need
the data for this long, but we do want to have some buffer to be able
to access historical data in the performance bar menu for a little
while.

We populate this information on each request for a user that has
access to the performance bar.

We also store a lot of information there including a detailed log of
queries, Gitaly and Redis calls. This makes this object quite big.