Fix SSH mirrors not working in FIPS mode
Previously on a FIPS system, adding a project mirror via SSH would
fail because the `ssh_data` gem attempted to generate the MD5 hash of
the server's public keys. To fix this, we return the SHA256 hash in
the controller and make the frontend prefer this when it is available.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/364562

Changelog: fixed