Allow job tokens to access internal packages
Packages are tied to the project visibility. If a project
is public, so is the package. If a project is internal, so
is the package. This updates permissions needed to pull
packages in those cases using the CI Job Token so that the
permissions match the behavior that already exists with
personal access tokens.

Related: https://gitlab.com/gitlab-org/gitlab/-/issues/333444
EE: true
Changelog: fixed