Correct validate epic_color_highlight FF in Epic mutations
The validation for the Epic `color` attribute was attempting to check
the feature flag status without passing in the group object to check the
feature flag against causing the attribute to always be removed unless
the feature flag was enabled globally.

The removal of the `color` attribute also happened at the wrong time.
The `color` attribute should be removed if the flag isn't enabled before
checking `args.empty?`.

Signed-off-by: Andrew Smith <espadav8@gmail.com>