Add the merge date to the Chain of Custody report
Prior to this change, the [Chain of Custody report][0] did not include
the date or time when a MR was merged.

Without this field, a user needed to take the report and
cross-reference it with a list of Git commits to identify the time a
merge occurred.

This was a time consuming and error prone process.

This lack of date information also made it difficult to filter reporte
rows by date ranges. See [issue 365041][1] for more details

With this commit, we now include a "Date Merged" column in the report

[0]:https://docs.gitlab.com/ee/user/compliance/compliance_report/#chain-of-custody-report
[1]:https://gitlab.com/gitlab-org/gitlab/-/issues/365041

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/91766#
EE: true