Pass full color object when set Epic color on new form
During a refactor of the handling of the colour objects with epics a
location was missed that only passes the colour value, not the full
colour object. This causes new epics to fail to be created with the
selected colour.

This was missed because the epic_form_spec wasn't updated at the same
time. This commit includes an update to the spec file to make sure that
colours objects are passed through the epic form.

Signed-off-by: Andrew Smith <espadav8@gmail.com>