Parse conflict lines with CRLF endings correctly
When lines have CRLF endings, GitLab is not able to parse them
correctly as we match the conflict markers after we delete the
newline character. That only deletes the LF and keep the CR.

To fix it, we also delete the CR before we match the conflict
markers.

Changelog: fixed