Add a test for when page size is retrieved from local storage
This adds a test to ensure the vulnerabilities query is only triggered
once with the correct `first` parameter when the page size is set in the
local storage. This covers a bug that was uncovered recently and that
results in the query being triggered twice when the component mounts
because it first tries to get the default page size, and immediately
makes another attempt with the page size retrieved by the
LocalStorageSync component.
This test voluntarily fails in this revision, the next commit will fix
the component to make this test pass.