Retrieve pageSize from local storage as soon as possible
This ensures the `VulnerabilityListGraphql` component retrieves the
user-selected page size from the local storage as soon as it initializes
its state. This ensures that the GraphQL query won't be fire twice: once
with the default page size, and a second time when the LocalStorageSync
emits an event with the actual size from the local storage.

Changelog: fixed
EE: true