Create index on security_findings asynchronously
We need an (uuid, id DESC) index on security_findings so that our new
Vulnerabilities::Finding creation service object doesn't timeout.

The timeout can occur when trying to fetch a security_findings rows via
an UUID because we have a UNIQUE index on (uuid, scan_id) which models
the possibility of finding the same exact vulnerability in subsequent
scans.

Changelog: added