Add sbom_occurrences table and model
This adds the `sbom_occurrences` table which is part three of a
three-phase implementation for adding new models to support the tracking
of Software Bill of Materials (SBoM) components in GitLab.
`sbom_occurrences` tracks each new occurence of a component, the commit
which it originated from, and its relation to other SBoM tables.

Changelog: added