Properly fail LDAP logins if GitLab user not persisted
Previously if restricted e-mail domains
(https://gitlab.com/gitlab-org/gitlab/-/issues/366269) or password
limit restrictions were set
(https://gitlab.com/gitlab-org/gitlab/-/issues/22399), newly-signed in
LDAP users would see an obscure failure message:

```
Could not authenticate you from Ldapmain because "Undefined method
`provider' for nil:nilclass".
```

This was failing in `Gitlab::Auth::Ldap::Access.open` because
`user.ldap_identity` was `nil`. The application settings would cause
validations of newly-signed in LDAP users to fail, preventing the user
from persisting to the database.

The OmniAuth callback controller attempted to detect this by calling
`auth_user.valid_sign_in?`, but this was failing because
`Gitlab::Auth::LdapUser#valid_sign_in?` did not properly check that
the user entry was persisted to the database before it contacted the
LDAP server.

To avoid this obscure error, we just need to swap the order of the
check: if the user entry is valid and persisted, then we can safely
contact the LDAP server. If the user entry is not valid, an "Access
Denied for your LDAP account" message will be shown to the user, and a
message in the application log will list the reason why it failed.

Changelog: fixed