Fix deploy keys breaking protected branch dropdown in FIPS mode
In FIPS mode, if a deploy key were present, clicking on the "Allowed
to push" dropdown would fail with a "Failed to load groups, users and
deploy keys" message. This occurred because the JavaScript attempted
to use the `fingerprint` attribute of the key.  However, on a FIPS
system, the MD5 fingerprint is not available, and the `null` value
breaks the dropdown.

To fix this, we use the `fingerprint_sha256` attribute instead. This
commit also puts the SHA256 fingerprint first in the HTML views to
avoid confusion.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/364562

Changelog: fixed