Fix flaky repository settings spec
The test was failing intermittently because the test would input the
URL and submit the data faster than the JavaScript was able to update
the internal field. This caused a blank URL to be submitted on
occasion. To make this spec more reliable, we now ensure the hidden
URL field is updated with the right value before clicking the "Save
changes" button.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/337394