Fix flaky legacy pipelines spec
The test was failing frequently because the Ajax request to play the
manual job was running after the `wait_for_requests`, so the backend
state of the job remained at `scheduled` instead of `pending`.

To reduce the chance of this happening, click the pipeline status
graph and look for a pending icon. Checking the state of the backend
shouldn't be required, but we include it just in case.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/367483