Ensure all Loose Foreign Key triggers have YML definitions
This is to prevent a previous issue we ran into where a loose foreign
key definition was removed before removing the trigger. This causes a
build up of loose foreign key deleted records which are never cleaned
up.

This MR adds a spec that looks at all the triggers and ensures that they
all have at least 1 corresponding YML definition that will handle the
deleted records. It also fixes the ordering in the documentation about
how to remove loose foreign keys. Technically they can all be in a
single MR but at least having the ordering right in case people want a
separate MR is a good idea.