Doc: Puma: Do not imply single mode has data integrity concerns
Although there are limitations and potentially stability and performance
concerns running puma in single mode has no data integrity problems as
far as is known. The current message raises concern with the way it is
phrased, so change the phrasing to avoid this implication.

See also the originating discussion for details:
https://gitlab.com/groups/gitlab-org/-/epics/5303#note_1016292201