Remove seed_helper from Gitlab::Git::Blame specs
SeedHelper uses git to directly insert repositories into git. Soon we
will be using the praefect database in tests and when this happens we
will not be able to directly manipulate repositories in gitaly without
praefects knowledge.

The repository used by the project factory has slighty different files
than the repository used by SeedHelper. It does not affect the function
of the tests.