Change feature flag check for hierarchy
When listing parent/children for a work item, we check only
generic work_item feature flag. This is little bit confusing (as for
creating/updating hiearchy widget we still use work_item_hiearchy FF),
but the reason is that parent/child relationship can be created by
"create from task" functionality and we still need to be able to list
parent for the tasks created this way.