Add second event handler oninput to help mitigate the issue
In extreme cases when the user clicks into the search header
too fast and starts typing no dropdown is shown. This is due to
the focus handler not being attached at the time of focus.
Adding a second event handler for input will catch when the user
starts typing.

Changelog: fixed