Take lease in UpdateProjectStatisticsWorker
Recent changes made to how we calculate repository size has made this a
much more expensive operation. Previously we used a naive `du -sk` call,
which can give inconsistent results depending on how well-packed the
repository is. Swapping to `git rev-list --all --objects --disk-usage`
gives us much more accurate results, but is orders of magnitude more
expensive to run.

To mitigate the increased load generated by this change, we want to
ensure we are limiting refreshes on project statistics to every 15
minutes. Currently we taken an exclusive lease in `ProjectCacheWorker`
to limit concurrency, but after 15 minutes this will trigger a
`UpdateProjectStatisticsWorker` call that does _not_ take a lease. As a
result, a subsequent push 15 minutes later will regenerate statistics
at the same time that the delayed job scheduled by the original push is
running.

Let's take a lease in `UpdateProjectStatisticsWorker` as well to ensure
that we're not running duplicate jobs.

Changelog: fixed