Update default value for project_import_level
This migration changes the default value for the
project_import_level column on the namespace_settings
table. It is needed because the initial default was not
mapped to the member.access_level values of a group (which
correspond to owner, maintiner, developer etc).

Changelog: fixed

Fix structure.sql schema file