Drop post migration test table
This table was introduced as part of
https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/585, to test the
execution of the new post migration (PDM) pipeline. With the testing and
rollout of the PDM being close to an end, we need to remove these test
table as well.

Even though this table was introduced in 15.2, it was never referenced
in the application codebase, so it's safe to remove it in this release
as well.

Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2352

Changelog: other