Add interpolation support for web-hook URLs
Prior to this change, URL variables were supported, but not used
during web-hook execution.

This change ensures that when we execute a web-hook, we take its
variables into account. Validation is added to ensure that web-hooks
are valid during execution.

Although new validations are added, we have not added any UI to create
variables, or started using interpolation yet, so existing records
should not be affected.