Support Project Avatar removal in REST API
This change set adds support to remove project avatars using the REST
API, much like it's already possible for the [Topics
API](https://docs.gitlab.com/ee/api/topics.html#remove-a-topic-avatar).

To support this I've changed the `POST | PUT /projects` endpoint to be
handled by workhorse. Can someone with more experience and the big
picture verify this particular part of the change? Is this something
which makes sense? Is it implemented correctly?

If this is the case and all looks good with the Projects API here, I'll
also implement similar changes to the other `Avatarable` endpoints, like
Groups and Users.

Changelog: added
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92604