Fix DNS stubbing not allowing all local networks
In our test helpers, we stub all DNS requests and only allow local
addresses to resolve via regexp. However, previously the regexp only
allowed the 172.16.x.x block, but the full private range is:

172.16.0.0 - 172.31.255.255.

In CI, the `postgres` service ended up in the 172.17.0.0 range. If this
IP were ever used, this change would enable the specs to
work. However, we still need to allow the original hostname.