Stop setting start_repository when it is the target repository
Always setting start_repository even when it is the same as the target
repository means that gitaly and praefect have to try and detect the
case where these two repos are identical in order to determine if we are
creating a new branch or not. This is not possible for praefect.