Fix move deploy keys during project imports in FIPS mode
When a project is imported, it may overwrite a project via
`OverwriteProjectService`, which then calls
`MoveDeployKeysProjectsService` to move deploy keys. Previously in
FIPS mode the deploy keys would fail to move over because the query
relied on a join with MD5 fingerprints, which aren't present in FIPS.

To fix this, we use SHA256 fingerprints instead. Since deploy keys may
be deleted, we ensure that SHA256 fingerprints exist for all the keys
in question.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/364562

Changelog: fixed