Use factory default deploy keys in git_access_spec.rb
Previously the spec created 2048-bit RSA keys, but these don't meet
the minimum 3072-bit requirement in FIPS system so the test was
failing on FIPS systems. Since the tests are agnostic to the number of
bits actually needed, we let the factory take care of the minimum
value needed for FIPS and non-FIPS systems.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/364562