Fix epic transfer when transfering a project
When transferring a project, we load all epics for issues in the
project and re-create the epic in the new group.

To find all the required epics, we join the issues from the project.
There was however a bug in this query as it would return the same epic
multiple time if an epic has multiple related issues.

To fix this, we're now use a `DISTINCT` on the query to only get each
epic once.

Changelog: fixed
EE: true