Verify correctness of ProtectedBranch cache
* Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/366724
* Follow-up for:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92922
* Feature flag: https://gitlab.com/gitlab-org/gitlab/-/issues/368279

**Problem**

Previous MR added a new implementation of ProtectedBranch cache. We want
to verify if it works correctly.

**Solution**

* Add `dry_run` option to ProtectedBranch cache
* Add feature flag to control cache flow

New cache implementation with `dry_run: true` will create new cache
structures but it will still return non-cached values to the user. If we
have an cache inconsistency then we will log an error message.