Drop /api/v4 suffix in GitLab OmniAuth documentation
With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/90866, the
`/api/v4` suffix is no longer needed in GitLab 15.2. It is a bit
confusing why this suffix was ever needed in the first place, but
configurations that use it will continue to function for now. The
upstream `omniauth-gitlab` gem has already dropped support for the
suffix entirely.

Changelog: changed