Fix notes count on MR to not increment when notes are loading
Before this fix, when there are multiple pages of discussions,
it's noticeable that the notes count on the MR page will increment
while discussions are getting loaded.

In this fix, we only update the count when all the pages of
discussions are done loading.

This is behind `paginated_mr_discussions` feature flag.