Skip repository disconnect if project is pending delete
We have seen Gitaly server saturation incidents where `git fsck` uses
a lot of memory checking the connectivity of a repository that is
about to be deleted. This is a side effect of our pool repository
application logic. This commit updates the pool repository application
logic to skip the work that causes the fsck.

https://gitlab.com/gitlab-org/gitlab/-/issues/368463

Changelog: performance