Disable rate limiting when transferring epics
When a project is transferred to a group, new epics may be created to
associate with the existing issues. However, epic creation is rate
limited. If there are too many epics, this would block the project
transfer.

To avoid this issue, disable rate limiting for transfers.

Relates to
https://gitlab.com/gitlab-org/gitlab/-/issues/368636

Changelog: fixed
EE: true