Remove duplicated false positives alert
We recently added the false positives alert to the shared
`vulnerability-details` component but the alert is still rendered in the
parent component (`vulnerability`). This causes the alert to show up
twice on the vulnerability details page. This removes the alert from the
page to only rely on the shared component's.

Changelog: fixed
EE: true