Remove RequirementsManagement::CreateRequirementService
Part of the process of turning requirements into
work items. Now that all requirements are work items
on database we only need the old objects to keep iids in sync.

CreateRequirementService is no longer needed, we can create
the requirement object that will be used as a proxy inside
a transaction on Issues::CreateService.

Changelog: other
EE: true