Update Markdown Serializer Spec Expectation
This is for handling the edge case of a markdown link ending with a
closing angular bracket like: `https://www.gitlab.com>`

Before the test asserted that it equals `https://www.gitlab.com>`, while
the markdown serializer actually serializes it to
`[https://www.gitlab.com>](https://www.gitlab.com%3E)`. This behavior
(which arguably seems to be correct) was confirmed in Firefox.

Seems like something underlying in JSDOMs URL implementation fixed this.