Allow Sbom::Occurence.component_version to be nullable
`.components[].version` is an optional field in CycloneDX 1.4. Because
of this, we need to account for the fact that the component version may
not be present. This change allows for
`sbom_occurrences.component_version_id` to be nullable, and instead
requires that a `component_id` must be present. This allows us to
continue to maintain a component list even if version information for
the component is not available.

Changelog: changed