Check group last owner before assigning default membership role
SAML Group Sync does not check if a member is the last owner
before assigning the default membership. In some cases this
led to groups being without an owner. To avoid this the worker
now checks for last owner before changing the member to default
membership role.

Changelog: fixed
EE: true