Fix RescheduleBackfillImportedIssueSearchData migration
In GitLab 14.9,
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/71913 introduced
a migration to backfill the issue search data.

In GitLab 15.2,
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/91979
rescheduled this migration. However, if between the two versions the
maximum `id` of the `issues` table decreased below the `max_value` in
the `batched_background_migrations` table for the 14.9 migration by an
issue deletion, this new migration would fail due to a database
constraint being violated.

To fix this, we need to take the minimum of the `SELECT MAX(id) FROM
issues` and the `max_value` from the table.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/368631

Changelog: fixed