Don't suggest caching of .venv
The option of caching whole .venvs seems attractive, because it represents the
end-result of an installation process and would therefore entail the least
rework between jobs. Unfortunately, venvs cannot be reliably relocated from one
directory to another, and this is what happens implicitly when using caches
across runners.

pypa's virtualenv is explicit about this; venv is not (but relocating still
doesn't work)
https://virtualenv.pypa.io/en/legacy/userguide.html#making-environments-relocatable

> The --relocatable option currently has a number of issues, and is not
> guaranteed to work in all circumstances. It is possible that the option will be
> deprecated in a future version of virtualenv.