Make ssh_signatures.key_id nullable
A signature may not have an associated key in GitLab, in which case we
use the `unknown_key` verification status. Currently, `ssh_signatures`
have a not null constraint on the `key_id`. This is not correct, because
it doesn't allow us to store signatures with the unknown key status.

Changelog: changed