Fix CI artifact sizes not logged for some runner endpoints
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/88140 may have
quietly broken the logging for the
`GET `api/:version/jobs/:id/artifacts` endpoint. Since
`log_artifact_size` was previously defined as a runner API helper and
a general API helper, the runner API version won. This caused the lazy
attribute loader to discard the `artifact` since it was a
`JobArtifactUploader` type instead of `Ci::JobArtifact`.

We can fix this by renaming the method to `log_artifact_file_size` and
adding a test to ensure the right type is called.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/369033

Changelog: fixed