Fix cross-join in old migration populate_status_column_of_security_scans
This migration was added in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/65201 which was in
14.3. As such this migration must have already been run before upgrading
to this version of GitLab.

We cannot remove or no-op this migration since it makes schema changes
but we remove the explicit `LOCK` statements which were only there for
performance reasons. The migration works fine and updates the schema the
same without them. They only served to reduce the risk of lock timeouts
on busy tables.