Update Audit Events Guide documentation with I18N information
During this [MR][0], where we were adding a new Audit Event, it was
mentioned by a reviewer that perhaps the `:message` attribute should
be translated via `I18N`.

In [this thread][1], it was determined that the `:message` attribute is
intentionally not translated.

We decided it would be a good idea to document the reasons for this to
help out in future Audit Event MRs

[0]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92068
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92068#note_1036881029

closes: https://gitlab.com/gitlab-org/gitlab/-/issues/368780