Execute side effects of merge request approval async
Before this, `MergeRequests::ApprovalService` will execute a
lot of SQL queries. When called with other services in a single
web request, that can lead 100+ SQL queries in a single queries.
That exceeds our query limit of 100.

This moves the side effects of approvals to be performed async
via Sidekiq. As a result, this reduces the number of SQL queries
executed in the service since it's offloaded to Sidekiq.

`Gitlab::EventStore` is utilized to make it easier in the
future to add side effects and opens the possiblity to refactor
the `MergeRequests::AfterApprovalWorker` to be split into smaller
subscribers.

This is behind `async_after_approval` feature flag.