GraphQL docs: Show descriptions for `:alpha` items
Previously the original descriptions were not displayed for "inline"
docs. This was because marking items as "alpha" leverages the our
deprecations, and for deprecated item we only display the deprecation
notice and not the original item description.

For alpha fields, we now display the warning of its instability, but do
also display the original description too.

Changelog: added