Fix OpenAPI viewer for dark mode
The OpenAPI viewer is not optimized for dark mode and is therefore
unreadable when dark mode is enabled. SwaggerUI currently does not
offer dark mode support as well. Using the CSS `filter` property
to invert the text color at least makes the text readable again when
dark mode is enabled.

Changelog: fixed