Clarify responsibilities for splitting large merge requests
Do not repeat the same responsibility for maintainers. We have feedback
that there is perceived and real weight of responsibilities so we want
to be very clear with responsibilities so that existing maintainers are
not burnt out, and new maintainers are not deterred.

Notwithstanding, this _does not_ prevent anyone from stepping up (see
GitLab's collaboration value).