Docs: Add section for paginated_mr_discussions feature flag
The `paginated_mr_discussions` feature flag is now enabled by
default on GitLab.com and self-managed installs (starting in 15.3).

This adds information about it on the docs so users will be able
to know that it's behind a feature flag and can be disabled in
case they experience some issues related to it.