Show better errors from GraphQL on Releases
A generic "Something went wrong..." error makes it hard for users to
understand how to fix any issues in the form. Instead, we should utilize
the good error messages provided by the various release services.

We also add a clarifying message for when a user attempts to create a
release with a protect tag when they are not allowed.

Changelog: added