Deprecate GraphQL `feature_flag` argument
This first iteration of
https://gitlab.com/gitlab-org/gitlab/-/issues/369202 documents the
`feature_flag` property as being deprecated and that it should not be
used.

It renames the argument so that it's clear it's not to be used for new
items.

It also updates the developer docs to include advice for using flags
with schema items besides fields, and where the flag is used to toggle
behavior.