Update merge request widget state actions
To make the states for the merge request better
match the new framework from the extensions, this makes the action
buttons more consistent with this.

For all states the action buttons have been changed to tertiary confirm
variant and right aligned.
When the screen size gets smaller the actions move into a dropdown.

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/368037