Fix issue description list item reordering indentation
Fix various styling issues by removing `list-style-position: inside`.
We extend the width of the drag icon so that it always overlaps
the `li` element so users can move from the `li` element to the
drag icon without a gap (a gap would make the drag icon disappear).

The tradeoff is that users can no longer make the drag icon
visible by hovering over the left of the list item marker.

Changelog: fixed