Switch back to license_finder and update to 7.0.1
In MR 74881, we needed to fork license_finder
as gitlab-license_finder as its strong dependencies to thor
(~> 1.0.1). With the release of the original license_finder 7.0.0,
the dependencies to thor was relaxed (~> 1.2), so we do not have to
use the fork.

The dependency problem was addressed in:
https://github.com/pivotal/LicenseFinder/pull/886
instead of https://github.com/pivotal/LicenseFinder/pull/816

Signed-off-by: Takuya Noguchi <takninnovationresearch@gmail.com>
Changelog: other