Revise text, completely change approach
The commit completely changes the approach of the original proposal.
It moves the per-version changes to the standard change format at
the top of the page. It shifts the information for Apache reverse
proxy users down to a new troubleshooting section, and completely
rephrases the information into the troubleshooting format.