Stop expanding file variables when sending to Runner
When file variables are referred in other variables, we should not
expand them directly because users expect that file variables are
expanded as their file paths, not their values. In order to expand them
as their file paths, we should send them to Runner unexpanded.

This change is behind a FF ci_stop_expanding_file_vars_for_runners