Changes how project export tarballs are uploaded to an external website
When requesting a project to be exported, the user can request the
export tarball to be uploaded to an external website, and in the case
object store is enabled, the upload uses `GitLab::HttpIO#read` to stream
read the file from object store. However, `GitLab::HttpIO#read` doesn't
perform well for large files as it makes several HTTP requests that
read small chunks of the file.

This change introduces a different method to stream files from object
store that only establishes one HTTP connection with the object store
and stream the file from the socket connection.