Show inherited permission in protected environment
When in a project's protected environments page, we want to show groups
that are able to deploy to protected environments. These permissions are
inherited by the project's parent group(s). A simple link to the group
is shown.

If the group cannot be loaded, a simple error message is displayed
within the popover.

Changelog: added