Extend gitlab:gitaly:check task with readiness verification
To verify Praefect is ready to server the requests
'gitlab:gitaly:check' task was extended to call 'ReadinessCheck' RPC.
The RPC was introduced in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/4674 and allows
to run 'praefect -config conf.toml check' sub-command from the remote.
If any check on remote fails it will be reported by the task. The
checks are executed only if Praefect is used, in case when setup has
no Praefect the RPC call to remote (in that case it would be Gitaly)
will always return no errors if the service is reachable.

Closes: https://gitlab.com/gitlab-org/gitlab/-/issues/348174