Fix calculation of averages for instrumentation metrics
This fixes the calculation of averages in Gitlab::Database::BatchCount
by introducing a new counter class that is specific to calculation of
average in batches. The new class works under the hood by calculating
the sum of all batches first, then dividing the result by the total
number of records that has non `null` values for that column.