Convert Query Analyzer feature flags to type: ops
These feature flags control "Query Analyzers". Query analyzers are
middleware that parse every SQL query sent to the database in Rails and
try to analyze and detect problems with the queries. These were needed
as part of CI decomposition to detect queries going to the wrong
database (ie. using the wrong model or connection).

These feature flags need to be set to `0.01%` permanently (or at least
never set to 100%) as they sample how frequently we parse queries in
production to detect problems. Parsing queries uses more CPU so we only
sample a subset of queries.

You can read more about these query analyzers in the merge requests that
introduced them:

1. https://gitlab.com/gitlab-org/gitlab/-/merge_requests/74177
2. https://gitlab.com/gitlab-org/gitlab/-/merge_requests/73839