Make the build hooks worker idempotent
Prior to this change, the payload was computed in the sidekiq worker
which may have different data due to replication lag.

This change computes the payload when we _enqueue_ the job, and skips
the use of the build-hooks worker, which is retained for backwards
compatibility, but is otherwise redundant.

Added: Adds execute_build_hooks_inline feature-flag

This adds a feature flag around the inline execution of
job hooks.