Update two_factor_authentication.md
The process has been updated from Support. Only 30 support contacts are allowed per customer. A SaaS user who needs a password reset should contact a named Support Contact. If a user were to open a 2FA reset request via a ticket, it would be automatically closed. For example see ticket 314363.

Message given to user if they did open a ticket would be:
GitLab SupportOps Bot

Greetings,

You're not listed as a support contact for your organization and this ticket will be closed. Please contact an administrator within your organization to either list you as an authorized contact or to raise a support request on your behalf.

We apologize for any inconvenience this might cause.

GitLab Support