Deprecate confidential attribute for notes
This deprecates the `confidential` parameter for the Notes API in favour
of `internal`. This affects the GraphQL API, REST API, and internal
Notes creation API.

We also expose the `internal` attribute now for a `Note` and the
behaviour of `internal` and `confidential` is the same.

In the case of submitting a request to create a note with
`internal: true`, and `confidential: false`, `internal` overrules the
`confidential` parameter.

In a follow-up work, the remaining mentions of `confidential` in the
Services, Models, and Database column will be renamed.

Changelog: deprecated