Add Gitlab::Redis::ALL_CLASSES constant
This makes the new constant a single source of truth for a new Redis
instance.

This makes sure we only need to add a new subclass of
`Gitlab::Redis::Wrapper` to define a new redis instance and add it to
this constant.

This will automatically add the instrumentation for the new redis
instance and add health checks for it to the readiness & liveliness
endpoints.

Doing so will also automatically generate the right spec helpers for
clearing the instance when needed.